requirements.txt
feedparser==6.0.11
requests==2.32.3
PyYAML==6.0.2
schedule==1.2.2
